Astrophysical Source Code Library (ASCL) entries for a certain year and month for submission to the Astrophysics Data System (ADS)
ascl.php-out.exe is a program to generate a text file of Astrophysical
Source Code Library (ASCL) entries for a certain year and month for
submission to the Astrophysics Data System (ADS), as well as an xls file
of fields for the entries.
The source code for the dynamic report of ASCL code entries is
downloaded from (http://asterisk.apod.com/ascl.php) and saved to a work
file. The file is then parsed and searched for the date queried, and the
entry fields are stored in a list.
The list of entries is then sorted. Then the fields for the entries are
parsed and html tags are removed to reveal the human readable fields.
The entries are then written to the output files and the program
execution ends.

The ADS takes the information from the file generated and then crates
standard ADS entries for the codes. The ASCL is treated as a
publication, and each code is a paper or article within the publication.
The combination of a unique ID for each code entry and the record in the
ADS allows codes to be cited independent of a published paper. The
benefits of this system are that being able to track citations to a code
could potentially help a coder build a record of publication and use for
his/her code, and that there is transparency in the methods used, to
know what codes were used (because they are cited) and being able to
look at them.